+++ R/grpc-package.R
@@ -1,2 +1,3 @@
 #' @useDynLib grpc, .registration = TRUE
+#' @importFrom methods as
 NULL